backend/requirements.txt
flask
flask-cors
pymupdf
faiss-cpu
sentence-transformers
ollama
numpy
